Added ko.dependencyDetection.preserveRepeaters
Nested computeds and subscriptions are disposed by default when a parent
computed or subscription reevaluates.  Call
preserveRepeaters(someCallback) to break the nesting relationship
between the computeds and subscriptions currently evaluating and those
contained in someCallback. Useful when you intend to cache the generated
computeds or when you really want your subscriptions to multiply
quadratically, for some reason.